playground/requirements_analytics.txt
# BigQuery Analytics Requirements
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0
google-cloud-bigquery==3.13.0
google-auth==2.23.4
